Refactor auth routes into blueprint
