+++ website/templates/index.html
@@ -1,5 +1,5 @@
 {% extends 'base.html' %}
 {% block content %}
 <h1>Generador de Horarios</h1>
-<p><a href="{{ url_for('login') }}" class="btn btn-primary">Ingresar</a></p>
+<p><a href="{{ url_for('auth.login') }}" class="btn btn-primary">Ingresar</a></p>
 {% endblock %}
