+++ website/templates/subscribe_success.html
@@ -9,7 +9,7 @@
   {% if session.get('user') %}
     <a href="{{ url_for('generador') }}" class="btn btn-primary mt-3">Ir al generador</a>
   {% else %}
-    <a href="{{ url_for('login') }}" class="btn btn-primary mt-3">Iniciar sesión</a>
+    <a href="{{ url_for('auth.login') }}" class="btn btn-primary mt-3">Iniciar sesión</a>
   {% endif %}
   <p class="mt-4"><a href="{{ url_for('landing') }}">Volver al inicio</a></p>
 </div>
